feat(db): update conversations and private messages schema

+++ backend/src/main/resources/db/migration/V11__privates_messages.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS private_messages (
     id              bigserial primary key,
     conversation_id bigint not null references conversations(id) on delete cascade,
-    sender_id       bigint not null references users(id) on delete cascade,
+    sender_id       integer not null references users(id) on delete cascade,
     content         text not null,
     created_at      timestamp DEFAULT CURRENT_TIMESTAMP,
     updated_at      timestamp DEFAULT CURRENT_TIMESTAMP
